requirements.txt
# Web Framework
fastapi==0.128.0
uvicorn==0.40.0
jinja2==3.1.2
aiofiles==23.2.1

# AI & LLM Orchestration
langchain==1.2.3
langchain-community==0.4.1
langchain-groq==1.1.1
langgraph==1.0.5
langsmith==0.6.2



# Vector Database & Embeddings
chromadb==1.5.2
sentence-transformers==5.2.0
huggingface-hub==0.36.0

# Python Scripts / Data loading
pandas
numpy

# Data Validation
pydantic==2.12.5

# Environment Variables
python-dotenv==1.2.1


# Database
psycopg2-binary==2.9.11